Split under the 500-line cap: decode-rows arm and its tests into sibling modules; attention diag helpers into trait_impl/diag.rs
hyper_connection_lowrank.rs 532 -> 392: the decode-rows arm (gate, shape contract, hc_pre_rows) moves to ops/hyper_connection_lowrank_rows.rs. hyper_connection_lowrank_tests.rs 686 -> 451: the rows parity tests and the DRAM-true microbench move to ops/hyper_connection_lowrank_rows_tests.rs; the shared Fixture and helpers open to the sibling as pub(super). qwen3_attention/trait_impl.rs 505 -> 434: diag_norm / diag_norm_f32 move verbatim to trait_impl/diag.rs with re-exports, so the V4 diagnostic call sites are unchanged. No behavior change; the six HC parity/microbench tests pass on the GPU fixture with --test-threads=1.
