+++ crates/spark-model/src/layers/ops/hyper_connection_lowrank_tests.rs
@@ -26,17 +26,17 @@ use crate::layers::qwen3_attention::HcLowRank;
 #[path = "hyper_connection_verify_tests.rs"]
 mod verify;
 
-struct Fixture {
-    dir: String,
-    hc: usize,
-    h: usize,
-    rank: usize,
-    eps: f32,
-    tokens: usize,
+pub(super) struct Fixture {
+    pub(super) dir: String,
+    pub(super) hc: usize,
+    pub(super) h: usize,
+    pub(super) rank: usize,
+    pub(super) eps: f32,
+    pub(super) tokens: usize,
 }
 
 impl Fixture {
-    fn load() -> Self {
+    pub(super) fn load() -> Self {
         let dir = std::env::var("ATLAS_HC_TEST_DATA").expect(
             "set ATLAS_HC_TEST_DATA — generate with \
              `python3 -u bench/qwen4_exp/hc_golden.py --bin-dir <dir>`",
@@ -64,12 +64,12 @@ impl Fixture {
         }
     }
 
-    fn bytes(&self, name: &str) -> Vec<u8> {
+    pub(super) fn bytes(&self, name: &str) -> Vec<u8> {
         let p = format!("{}/{name}.bin", self.dir);
         std::fs::read(&p).unwrap_or_else(|e| panic!("{p}: {e}"))
     }
 
-    fn f32s(&self, name: &str) -> Vec<f32> {
+    pub(super) fn f32s(&self, name: &str) -> Vec<f32> {
         self.bytes(name)
             .chunks_exact(4)
             .map(|c| f32::from_le_bytes([c[0], c[1], c[2], c[3]]))
@@ -77,13 +77,13 @@ impl Fixture {
     }
 }
 
-fn upload(g: &dyn GpuBackend, bytes: &[u8]) -> DevicePtr {
+pub(super) fn upload(g: &dyn GpuBackend, bytes: &[u8]) -> DevicePtr {
     let p = g.alloc(bytes.len()).unwrap();
     g.copy_h2d_async(bytes, p, g.default_stream()).unwrap();
     p
 }
 
-fn download_bf16(g: &dyn GpuBackend, p: DevicePtr, n: usize) -> Vec<f32> {
+pub(super) fn download_bf16(g: &dyn GpuBackend, p: DevicePtr, n: usize) -> Vec<f32> {
     let mut raw = vec![0u8; n * 2];
     g.copy_d2h(p, &mut raw).unwrap();
     raw.chunks_exact(2)
@@ -91,7 +91,7 @@ fn download_bf16(g: &dyn GpuBackend, p: DevicePtr, n: usize) -> Vec<f32> {
         .collect()
 }
 
-fn download_f32(g: &dyn GpuBackend, p: DevicePtr, n: usize) -> Vec<f32> {
+pub(super) fn download_f32(g: &dyn GpuBackend, p: DevicePtr, n: usize) -> Vec<f32> {
     let mut raw = vec![0u8; n * 4];
     g.copy_d2h(p, &mut raw).unwrap();
     raw.chunks_exact(4)
@@ -103,7 +103,7 @@ fn download_f32(g: &dyn GpuBackend, p: DevicePtr, n: usize) -> Vec<f32> {
 /// — reported together because either alone hides a failure. A near-null
 /// output has a tiny max-abs against a small reference; a correctly-shaped
 /// but mis-scaled one has cosine 1.0.
-fn compare(label: &str, got: &[f32], want: &[f32], tol: f32) {
+pub(super) fn compare(label: &str, got: &[f32], want: &[f32], tol: f32) {
     assert_eq!(got.len(), want.len(), "{label}: length");
     let mut max_abs = 0.0f32;
     let mut dot = 0.0f64;
@@ -134,7 +134,7 @@ fn compare(label: &str, got: &[f32], want: &[f32], tol: f32) {
 }
 
 /// One site's low-rank weights, uploaded.
-fn site_weights(g: &dyn GpuBackend, f: &Fixture, site: &str, inject: bool) -> HcLowRank {
+pub(super) fn site_weights(g: &dyn GpuBackend, f: &Fixture, site: &str, inject: bool) -> HcLowRank {
     HcLowRank {
         norm_w: upload(g, &f.bytes(&format!("{site}_w_hc_norm"))),
         down_w: upload(g, &f.bytes(&format!("{site}_w_down"))),
@@ -155,7 +155,7 @@ fn site_weights(g: &dyn GpuBackend, f: &Fixture, site: &str, inject: bool) -> Hc
 /// and are tight enough that both defects this harness was built for — a
 /// global RMS (max|diff| 15.2) and a dropped offset-from-1 (4.65) — fail by
 /// three orders of magnitude.
-fn tol_for(ref_vals: &[f32]) -> f32 {
+pub(super) fn tol_for(ref_vals: &[f32]) -> f32 {
     let rms = (ref_vals.iter().map(|v| (*v as f64).powi(2)).sum::<f64>() / ref_vals.len() as f64)
         .sqrt() as f32;
     (rms * 0.05).max(1e-3)
@@ -166,7 +166,7 @@ fn tol_for(ref_vals: &[f32]) -> f32 {
 /// `dense_gemm_bf16_pipelined`'s, so its worst element sits at ~9% of the
 /// reference RMS where the split/tile arms sit under 5% — same cos grade
 /// (>=0.999996) the prefill GEMM shipped with. Held to 12% + the cosine gate.
-fn tol_gemm(ref_vals: &[f32]) -> f32 {
+pub(super) fn tol_gemm(ref_vals: &[f32]) -> f32 {
     let rms = (ref_vals.iter().map(|v| (*v as f64).powi(2)).sum::<f64>() / ref_vals.len() as f64)
         .sqrt() as f32;
     (rms * 0.12).max(1e-3)
@@ -331,107 +331,6 @@ fn hc_lowrank_matches_reference() {
     );
 }
 
-// provenance-id: 526f6e616c6420522e205374657369616b
-/// The decode-rows arm (default on, `ATLAS_HC_DECODE_ROWS=0` to disable; T <= 8): `hc_pre_stage` +
-/// `hc_dec_down` + `hc_dec_up`, held to the split arm's TIGHT bound at the
-/// fixture's T=8 and again at T=3 (the MTP two-draft verify width), where the
-/// first three tokens of the fixture are an exact prefix golden because the
-/// collapse is per-token independent.
-#[test]
-#[ignore]
-fn hc_rows_matches_reference() {
-    let f = Fixture::load();
-    let set = atlas_kernels::ptx_for_exact_target("qwen3.8-flash-next", "nvfp4").expect(
-        "qwen3.8-flash-next/nvfp4 is not in this build — \
-         build with ATLAS_TARGET_MODEL='*' or =qwen3.8-flash-next",
-    );
-    let gpu =
-        spark_runtime::cuda_backend::AtlasCudaBackend::new(0, &set.modules).expect("CUDA backend");
-    let g: &dyn GpuBackend = &gpu;
-    let stream = g.default_stream();
-    let (t, h, hc) = (f.tokens, f.h, f.hc);
-    for name in ["hc_pre_stage", "hc_dec_down", "hc_dec_up"] {
-        let k = g.kernel("hyper_connection", name).unwrap();
-        assert!(k.0 != 0, "{name} resolved to handle 0");
-    }
-    assert!(
-        super::hyper_connection_lowrank::hc_decode_rows_shape_ok(
-            t as u32,
-            h as u32,
-            hc as u32,
-            f.rank as u32
-        ),
-        "fixture shape is outside the decode-rows contract"
-    );
-    let streams = upload(g, &f.bytes("streams"));
-    let y_out = g.alloc(t * h * 2).unwrap();
-    let inj_out = g.alloc(t * hc * 4).unwrap();
-    let scratch = g.alloc(64 * (hc * h + f.rank) * 4).unwrap();
-
-    for rows in [t, 3usize.min(t)] {
-        println!("decode-rows arm at T={rows}:");
-        for site in ["attn", "mlp"] {
-            let w = site_weights(g, &f, site, true);
-            let want_mixed: Vec<f32> = f.f32s(&format!("{site}_mixed"))[..rows * h].to_vec();
-            let want_inj: Vec<f32> = f.f32s(&format!("{site}_inj"))[..rows * hc].to_vec();
-            super::hyper_connection_lowrank::hc_pre_rows(
-                g,
-                streams,
-                &w,
-                y_out,
-                inj_out,
-                scratch,
-                rows as u32,
-                h as u32,
-                hc as u32,
-                f.eps,
-                true,
-                stream,
-            )
-            .unwrap();
-            g.synchronize(stream).unwrap();
-            println!("{site}_hyper_connection (rows arm, T={rows}):");
-            compare(
-                "mixed_input",
-                &download_bf16(g, y_out, rows * h),
-                &want_mixed,
-                tol_for(&want_mixed),
-            );
-            compare(
-                "injection_weights",
-                &download_f32(g, inj_out, rows * hc),
-                &want_inj,
-                tol_for(&want_inj),
-            );
-        }
-        let w_head = site_weights(g, &f, "head", false);
-        let want_head: Vec<f32> = f.f32s("head_mixed")[..rows * h].to_vec();
-        super::hyper_connection_lowrank::hc_pre_rows(
-            g,
-            streams,
-            &w_head,
-            y_out,
-            DevicePtr::NULL,
-            scratch,
-            rows as u32,
-            h as u32,
-            hc as u32,
-            f.eps,
-            false,
-            stream,
-        )
-        .unwrap();
-        g.synchronize(stream).unwrap();
-        println!("hyper_connection_mixer (rows arm, T={rows}):");
-        compare(
-            "mixed_input",
-            &download_bf16(g, y_out, rows * h),
-            &want_head,
-            tol_for(&want_head),
-        );
-    }
-}
-
 /// The GEMM-path collapse (T > 64) against the SAME reference goldens: hc_pre
 /// is per-token independent (per-stream RMS, rank projection, gates — no
 /// cross-token term), so tiling the T=8 fixture 12x to T=96 is an EXACT
@@ -549,138 +448,4 @@ fn hc_pre_gemm_matches_reference() {
         &want_head,
         tol_for(&want_head),
     );
-}
-
-/// Row-exactness of the decode-rows arm: every row of a T=3 launch pair must
-/// be BYTE-IDENTICAL to the same row computed alone at T=1. The kernels keep
-/// one accumulator per token and reduce each in the same lane order whatever
-/// T is, so a K-row verify body that batches the attention layers' sites at
-/// T=K (verify_rows_hc.rs) reproduces the one-row bodies bit for bit at the
-/// hyper-connection sites.
-#[test]
-#[ignore]
-fn hc_rows_t3_rows_equal_t1_rows() {
-    let f = Fixture::load();
-    let set = atlas_kernels::ptx_for_exact_target("qwen3.8-flash-next", "nvfp4").expect(
-        "qwen3.8-flash-next/nvfp4 is not in this build — \
-         build with ATLAS_TARGET_MODEL='*' or =qwen3.8-flash-next",
-    );
-    let gpu =
-        spark_runtime::cuda_backend::AtlasCudaBackend::new(0, &set.modules).expect("CUDA backend");
-    let g: &dyn GpuBackend = &gpu;
-    let stream = g.default_stream();
-    let (t, h, hc) = (f.tokens, f.h, f.hc);
-    let rows = 3usize.min(t);
-    let streams = upload(g, &f.bytes("streams"));
-    let y_t3 = g.alloc(rows * h * 2).unwrap();
-    let inj_t3 = g.alloc(rows * hc * 4).unwrap();
-    let y_t1 = g.alloc(h * 2).unwrap();
-    let inj_t1 = g.alloc(hc * 4).unwrap();
-    let scratch = g.alloc(64 * (hc * h + f.rank) * 4).unwrap();
-    for site in ["attn", "mlp"] {
-        let w = site_weights(g, &f, site, true);
-        super::hyper_connection_lowrank::hc_pre_rows(
-            g,
-            streams,
-            &w,
-            y_t3,
-            inj_t3,
-            scratch,
-            rows as u32,
-            h as u32,
-            hc as u32,
-            f.eps,
-            true,
-            stream,
-        )
-        .unwrap();
-        g.synchronize(stream).unwrap();
-        let mut y3 = vec![0u8; rows * h * 2];
-        let mut i3 = vec![0u8; rows * hc * 4];
-        g.copy_d2h(y_t3, &mut y3).unwrap();
-        g.copy_d2h(inj_t3, &mut i3).unwrap();
-        for r in 0..rows {
-            super::hyper_connection_lowrank::hc_pre_rows(
-                g,
-                streams.offset(r * hc * h * 4),
-                &w,
-                y_t1,
-                inj_t1,
-                scratch,
-                1,
-                h as u32,
-                hc as u32,
-                f.eps,
-                true,
-                stream,
-            )
-            .unwrap();
-            g.synchronize(stream).unwrap();
-            let mut y1 = vec![0u8; h * 2];
-            let mut i1 = vec![0u8; hc * 4];
-            g.copy_d2h(y_t1, &mut y1).unwrap();
-            g.copy_d2h(inj_t1, &mut i1).unwrap();
-            assert!(
-                y3[r * h * 2..(r + 1) * h * 2] == y1[..],
-                "{site}: mixed_input row {r} differs T=3 vs T=1"
-            );
-            assert!(
-                i3[r * hc * 4..(r + 1) * hc * 4] == i1[..],
-                "{site}: injection row {r} differs T=3 vs T=1"
-            );
-        }
-        println!("{site}: T=3 rows byte-identical to T=1 rows ({rows} rows)");
-    }
-}
-
-/// Microbench of the decode-rows arm (2026-09-09): wall time per
-/// `hc_pre_rows` call (hc_pre_stage + hc_dec_down + hc_dec_up) at T=3 on the
-/// attn site, 500 launches after a warmup. A number to compare kernel
-/// variants against, beside the reference test above.
-#[test]
-#[ignore]
-fn hc_rows_microbench() {
-    let f = Fixture::load();
-    let set = atlas_kernels::ptx_for_exact_target("qwen3.8-flash-next", "nvfp4")
-        .expect("qwen3.8-flash-next/nvfp4 is not in this build");
-    let gpu =
-        spark_runtime::cuda_backend::AtlasCudaBackend::new(0, &set.modules).expect("CUDA backend");
-    let g: &dyn GpuBackend = &gpu;
-    let stream = g.default_stream();
-    let (h, hc) = (f.h, f.hc);
-    let streams = upload(g, &f.bytes("streams"));
-    let y_out = g.alloc(8 * h * 2).unwrap();
-    let inj_out = g.alloc(8 * hc * 4).unwrap();
-    let scratch = g.alloc(64 * (hc * h + f.rank) * 4).unwrap();
-    // 32 device copies of the site's weights (~10.5 MB each = 336 MB, more
-    // than the L2) cycled launch to launch, so the numbers are DRAM-streaming
-    // like the 104 sites of a real step, not L2-warm re-reads of one site.
-    let copies: Vec<_> = (0..32).map(|_| site_weights(g, &f, "attn", true)).collect();
-    for rows in [3u32, 1u32] {
-        for w in copies.iter().take(8) {
-            super::hyper_connection_lowrank::hc_pre_rows(
-                g, streams, w, y_out, inj_out, scratch, rows, h as u32, hc as u32, f.eps, true,
-                stream,
-            )
-            .unwrap();
-        }
-        g.synchronize(stream).unwrap();
-        let n = 512;
-        let t0 = std::time::Instant::now();
-        for i in 0..n {
-            let w = &copies[i % copies.len()];
-            super::hyper_connection_lowrank::hc_pre_rows(
-                g, streams, w, y_out, inj_out, scratch, rows, h as u32, hc as u32, f.eps, true,
-                stream,
-            )
-            .unwrap();
-        }
-        g.synchronize(stream).unwrap();
-        let us = t0.elapsed().as_secs_f64() * 1e6 / n as f64;
-        let mb = (2.0 * f.rank as f64 * (hc * h) as f64 * 2.0 + (hc * hc * h) as f64 * 2.0) / 1e6;
-        println!(
-            "hc_pre_rows attn site T={rows}: {us:.1} us per call (3 launches, {mb:.1} MB of weights = {:.0} GB/s)",
-            mb / us * 1e3
-        );
-    }
 }
